+++ src/components/Product/Footer/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from 'next/image';
 import footerStyles from './Footer.module.scss';
-import productStyles from '../Product.module.scss';
+import productStyles from '../../../styles/pages/Product.module.scss';
 import { signIn } from 'next-auth/react';
 import { Button } from '../../Global';
 import ubcLogoFullBlack from '../../../../public/assets/UBC-logo-full-black.png';
